Added smart contract + updated README
Added the smart contract with a constructor to initialize a
collection of candidates, a validation method to determine if a
candidate exists, a setter method to cast a vote for a candidate,
and a getter method to get the number of votes that a candidate
has. Updated the README with instructions on how to compile the
smart contract.